requirements.txt
gTTS==2.3.1
numpy==1.24.2
openai==0.27.4
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==10.2.0
pydub==0.25.1
python_vlc==3.0.20123
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
SpeechRecognition==3.10.0
spotipy==2.23.0
tiktoken==0.3.3
pydub
pygame
